// packages/app/api/generated/models/ChatMessageResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ChatMessageResponse
 */
export interface ChatMessageResponse {
    /**
     * 
     * @type {string}
     * @memberof ChatMessageResponse
     */
    messageId?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatMessageResponse
     */
    senderId?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatMessageResponse
     */
    message?: string;
    /**
     * 
     * @type {number}
     * @memberof ChatMessageResponse
     */
    timestamp?: number;
}

/**
 * Check if a given object implements the ChatMessageResponse interface.
 */
export function instanceOfChatMessageResponse(value: object): value is ChatMessageResponse {
    return true;
}

export function ChatMessageResponseFromJSON(json: any): ChatMessageResponse {
    return ChatMessageResponseFromJSONTyped(json, false);
}

export function ChatMessageResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChatMessageResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'messageId': json['messageId'] == null ? undefined : json['messageId'],
        'senderId': json['senderId'] == null ? undefined : json['senderId'],
        'message': json['message'] == null ? undefined : json['message'],
        'timestamp': json['timestamp'] == null ? undefined : json['timestamp'],
    };
}

export function ChatMessageResponseToJSON(json: any): ChatMessageResponse {
    return ChatMessageResponseToJSONTyped(json, false);
}

export function ChatMessageResponseToJSONTyped(value?: ChatMessageResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'messageId': value['messageId'],
        'senderId': value['senderId'],
        'message': value['message'],
        'timestamp': value['timestamp'],
    };
}